Made the circle algorithm to work with steps rather than mm, almost there. A few more fixes and it should work!

+++ Inc/gcodeInterpreter.h
@@ -7,7 +7,7 @@ void drawLine(int32_t x0, int32_t y0, int32_t x1, int32_t y1);
 /*Rasterizes an arc into small line segements and runs bresenhams on each of those*/
 void drawArc(float x, float y, float I, float J, float direction);
 
-bool check(float x0, float y0, float x1, float y1);
-bool complete(float x0, float y0, float x1, float y1);
+bool check(int32_t x0, int32_t y0, int32_t x1, int32_t y1);
+bool complete(int32_t x0, int32_t y0, int32_t x1, int32_t y1);
 void arc(float x1mm, float y1mm, float I, float J);
-void nextStep(float x0, float y0, float x1, float y1, float r);
+void nextStep(int32_t x0, int32_t y0, int32_t x1, int32_t y1, int32_t r);
